requirements.txt
discord
pandas
matplotlib
python-dotenv